Broke solutionParser out into it's own file, refactored code and added comments
